Add files via upload
This C# code uses the Autolab Metrohm SDK. It sets Autolab to galvanostatic mode. It sets the offset current through the setpoint. It sets the amplitude of the FRA module. It changes the frequency of the FRA module and measures complex impedance. The EIS data are set as text files.